+++ client/views/posts/modules/post_info.js
@@ -1,4 +1,4 @@
-Template[getTemplate('postMeta')].helpers({
+Template[getTemplate('postInfo')].helpers({
   pointsUnitDisplayText: function(){
     return this.upvotes == 1 ? i18n.t('point') : i18n.t('points');
   },