+++ client/views/posts/posts_list_incoming.js
@@ -1,0 +1,5 @@
+Template[getTemplate('postsListIncoming')].events({
+  'click .show-new': function(e, instance) {
+    Session.set('listPopulatedAt', new Date());
+  }
+});