+++ packages/telescope-newsletter/package.js
@@ -1,6 +1,13 @@
-Package.describe({summary: "Telescope email digest package"});
+Package.describe({
+  summary: "Telescope email newsletter package",
+  version: '0.1.0',
+  name: "telescope-newsletter"
+});
 
-Npm.depends({later: "1.1.6"});
+Npm.depends({
+  "html-to-text": "0.1.0",
+  "later": "1.1.6"
+});
 
 Package.onUse(function (api) {
 
@@ -30,9 +37,9 @@ Package.onUse(function (api) {
   ], ['client', 'server']);
 
   api.add_files([
-    'lib/client/newsletter_banner.html',
-    'lib/client/newsletter_banner.js',
-    'lib/client/newsletter_banner.css'
+    'lib/client/templates/newsletter_banner.html',
+    'lib/client/templates/newsletter_banner.js',
+    'lib/client/templates/newsletter_banner.css'
   ], ['client']);
 
   api.add_files([