+++ History.md
@@ -1,4 +1,12 @@
+## v0.9.4 “UpdateScope”
+
 * Removed unneeded allow insert on Posts and Comments. 
+* Renaming `postMeta` template to `postInfo` to avoid ambiguity.
+* Fixing avatar code.
+* Adding update prompt package.
+* Upgrading to Meteor 0.9.2.
+* Do not remove deleted comments from database, use `isDeleted` flag instead.
+* Now showing "X new posts" instead of just displaying them. Thanks @dburles!
 
 ## v0.9.3 “DailyScope”